Move definitions of CommonViewType helpers
